+++ x/testutils/mocks/amqp_client.go
